Move submission flow components in to subdir
